tests/wpt/web-platform-tests/html/webappapis/scripting/events/uncompiled_event_handler_with_scripting_enabled.html

<!doctype html>
<meta charset="utf-8">
<title>Uncompiled event handler should return compiled handler with scripting enabled</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(function() {
    var invoked = false;
    document.addEventListener("click", function() {
      invoked = true;
    });
    var click = new MouseEvent("click");
    document.dispatchEvent(click);
    assert_equals(invoked, true);
  });
</script>
